refactor metrics tests
