+++ test/e2e/tests/metrics/mock-data.js
