+++ test/e2e/tests/metrics/permissions-approved.spec.js
@@ -6,8 +6,8 @@ const {
   openDapp,
   unlockWallet,
   getEventPayloads,
-} = require('../helpers');
-const FixtureBuilder = require('../fixture-builder');
+} = require('../../helpers');
+const FixtureBuilder = require('../../fixture-builder');
 
 /**
  * mocks the segment api multiple times for specific payloads that we expect to
