+++ test/e2e/tests/metrics/signature-approved.spec.js
@@ -9,8 +9,8 @@ const {
   getEventPayloads,
   clickSignOnSignatureConfirmation,
   validateContractDetails,
-} = require('../helpers');
-const FixtureBuilder = require('../fixture-builder');
+} = require('../../helpers');
+const FixtureBuilder = require('../../fixture-builder');
 
 /**
  * mocks the segment api multiple times for specific payloads that we expect to
