+++ test/e2e/tests/metrics/transaction-finalized.spec.js
@@ -7,8 +7,8 @@ const {
   getEventPayloads,
   assertInAnyOrder,
   logInWithBalanceValidation,
-} = require('../helpers');
-const FixtureBuilder = require('../fixture-builder');
+} = require('../../helpers');
+const FixtureBuilder = require('../../fixture-builder');
 
 /**
  * mocks the segment api multiple times for specific payloads that we expect to
